Moved GenerateBackupCodes into its own component to resolve build-time type constraint violation
